# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sito DNS Administrative API",
    "description": "Administrative REST API and OpenAPI 3.0 specification for the sito DNS server",
    "license": {
      "name": "GPL-3.0-only"
    },
    "version": "1.6.0"
  },
  "paths": {
...
    "/api/v1/querylog": {
      "get": {
        "tags": [
          "QueryLog"
        ],
        "operationId": "get_querylog",
        "parameters": [
          {
            "name": "client",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "domain",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "qtype",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Query logs retrieved with cursor pagination",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryLogPage"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "QueryLog"
        ],
        "operationId": "delete_querylog",
        "responses": {
          "200": {
            "description": "Query log cleared successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericMessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GenericMessageResponse": {
        "type": "object",
        "description": "Generic success message response.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 Problem Details response object.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Detailed explanation specific to this occurrence."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional URI reference identifying the specific occurrence of the problem."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary of the problem."
          },
          "type": {
            "type": "string",
            "description": "URI reference identifying the problem type."
          }
        }
      },
      "QueryLogEntry": {
        "type": "object",
        "description": "A single DNS query log record stored in SQLite or streamed over WebSockets.",
        "required": [
          "ts",
          "client_ip",
          "qname",
          "qtype",
          "verdict",
          "proto"
        ],
        "properties": {
          "client_ip": {
            "type": "string",
            "description": "Client IP address (or masked if anonymization is enabled)."
          },
          "client_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional human-readable client name."
          },
          "dnssec": {
            "type": [
              "string",
              "null"
            ],
            "description": "DNSSEC outcome: \"secure\" | \"insecure\" | \"bogus\"."
          },
          "elapsed_us": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Total elapsed processing duration in microseconds."
          },
          "id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "list_source": {
            "type": [
              "string",
              "null"
            ],
            "description": "Source list name that produced the rule match."
          },
          "proto": {
            "type": "string",
            "description": "Incoming protocol: \"udp\" | \"tcp\" | \"dot\" | \"doh\" | \"doq\" | \"doh3\"."
          },
          "qname": {
            "type": "string",
            "description": "Fully-qualified domain name queried (lowercased without trailing dot)."
          },
          "qtype": {
            "type": "integer",
            "format": "int32",
            "description": "Query record type (A=1, AAAA=28, HTTPS=65, etc.).",
            "minimum": 0
          },
          "rcode": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Response code (0=NoError, 3=NXDomain, 2=ServFail, etc.).",
            "minimum": 0
          },
          "rule": {
            "type": [
              "string",
              "null"
            ],
            "description": "Matched filter rule text if any."
          },
          "ts": {
            "type": "integer",
            "format": "int64",
            "description": "Timestamp in unix milliseconds."
          },
          "upstream": {
            "type": [
              "string",
              "null"
            ],
            "description": "Upstream resolver name or address that answered."
          },
          "verdict": {
            "type": "string",
            "description": "Filtering verdict: \"allowed\" | \"blocked\" | \"whitelisted\" | \"rewritten\" | \"stale\"."
          }
        }
      },
...
      "QueryLogPage": {
        "type": "object",
        "description": "Paginated page of query logs.",
        "required": [
          "entries"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryLogEntry"
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "total_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
    }
  },
...
}